requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astor==0.8.1
bce-python-sdk==0.9.67
certifi==2026.2.25
chardet==7.4.0.post1
charset-normalizer==3.4.6
click==8.3.1
colorlog==6.10.1
coverage==7.13.5
crc32c==2.8
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
decorator==5.2.1
easyocr==1.7.2
filelock==3.25.2
fsspec==2026.3.0
future==1.0.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
ImageIO==2.37.3
imagesize==2.0.0
iniconfig==2.3.0
Jinja2==3.1.6
lazy-loader==0.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modelscope==1.35.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt-einsum==3.3.0
packaging==26.0
paddleocr==3.4.0
paddlepaddle-gpu==2.6.2
paddlex==3.4.3
pandas==3.0.1
pillow==12.1.1
pluggy==1.6.0
prettytable==3.17.0
protobuf==7.34.1
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdfium2==5.6.0
pytesseract==0.3.13
pytest==9.0.2
pytest-cov==7.1.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2026.3.32
requests==2.33.0
rich==14.3.3
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tifffile==2026.3.3
tkinterdnd2==0.4.3
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.4.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
wcwidth==0.6.0
